backend/requirements.txt
# Flask y extensiones
Flask==2.3.3
flask-cors==4.0.0  #para frontend y backend puedan comunicarse

# Base de datos PostgreSQL
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23

# Variables de entorno
python-dotenv==1.0.0

# Autenticación JWT
PyJWT==2.8.0

# Utilidades
python-dateutil==2.8.2 
requests==2.31.0 #para hacer peticiones a la API

# Desarrollo (opcional)
flask-migrate==4.0.5 #para hacer migraciones de la base de datos
